refactor: rename route's page file from index.tsx to Page.tsx
