+++ meshapp/src/pages/SignUp/components/inputs/CustomCheckBox.tsx
@@ -1,5 +1,5 @@
 import { Controller, useFormContext } from "react-hook-form";
-import { IFormInput } from "../..";
+import { IFormInput } from "../../Page";
 import { Checkbox, FormControlLabel, Grid, Tooltip } from "@mui/material";
 import ErrorIcon from "@mui/icons-material/Error";
