+++ meshapp/src/pages/SignUp/components/inputs/CustomSelect.tsx
@@ -1,5 +1,5 @@
 import { Controller, useFormContext } from "react-hook-form";
-import { IFormInput } from "../..";
+import { IFormInput } from "../../Page";
 import {
   Box,
   FormControl,
